domains/auth/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.26.0
email-validator==2.1.0
prometheus-client==0.20.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
